+++ app/src/main/java/cz/yetanotherview/webcamviewer/app/drawer/NavigationDrawerCallbacks.java
@@ -1,5 +1,5 @@
 package cz.yetanotherview.webcamviewer.app.drawer;
 
 public interface NavigationDrawerCallbacks {
-    void onNavigationDrawerItemSelected(int position, long categoryId);
+    void onNavigationDrawerItemSelected(int position, int categoryId);
 }
